backend/requirements.txt
fastapi
uvicorn
pypdf2
langchain_google_genai
python-docx
python-dotenv
python-multipart